Check if the inputted user name is taken already
Already had this checking in place for the email field, but realized that
usernames also need to be unique. This is different from the real Twitter
(they seem to generate random chars to add to the end of a username that
is already taken, but I don't like that approach at all, and I think
usernames used to be unique on Twitter...).